requirements.txt
feedparser==6.0.10
python-telegram-bot==13.15